help

+++ main.cpp
@@ -1,12 +1,9 @@
 #include <iostream>
 #include "Banco.h"
-
 using namespace std;
-
 int main() {
 	Banco bancoUCC;
-	int opcion;
-
+	short opcion;
 	do {
 		cout << "--- Menu de Gestion Banco UCC ---" << endl;
 		cout << "1. Registrar Cliente" << endl;
@@ -18,7 +15,6 @@ int main() {
 		cout << "0. Salir" << endl;
 		cout << "Seleccione una opcion: ";
 		cin >> opcion;
-
 		switch (opcion) {
 		case 1:
 			bancoUCC.registrarCliente();
@@ -47,6 +43,5 @@ int main() {
 		}
 		cout << endl;
 	} while (opcion != 0);
-
 	return 0;
 }
